Update README.md
This update includes changes to the README.md to provide more detailed information about my current projects, learning progress, areas of collaboration, and how to reach out. It also includes fun and interesting facts related to my passions, such as technology and aviation, to give a more personal touch to my GitHub profile. The goal is to make this profile more engaging and informative for anyone looking to collaborate or learn more about my work.